Remove renderer.tsx and merge into default index.tsx
